solved timing issue on Windows for discrete sampling sources
