+++ multisensor_pipeline/tests/test_basic.py
@@ -245,53 +245,3 @@ class BaseTestCase(unittest.TestCase):
         self.assertEqual(len(list_sink), 10)
         self.assertEqual(sleep_trash_sink.counter, 10)
         self.assertEqual((end_time - start_time).seconds, 5)
-
-    def test_fixed_rate_source_under_load(self):
-
-        class TimestampListSink(BaseSink):
-
-            def __init__(self):
-                super(TimestampListSink, self).__init__()
-                self.timestamps = []
-
-            def on_update(self, frame: MSPDataFrame):
-                self.timestamps.append(frame.timestamp)
-
-        def _run_test_pipeline(samplerate: int, shape):
-            # define the modules
-            source = RandomArraySource(
-                shape=shape,
-                samplerate=samplerate,
-                max_count=samplerate
-            )
-            sink = TimestampListSink()
-
-            # add module to a pipeline...
-            pipeline = GraphPipeline()
-            pipeline.add(modules=[source, sink])
-            # ...and connect the modules
-            pipeline.connect(module=source, successor=sink)
-
-            pipeline.start()
-            sleep(1.5)
-            pipeline.stop()
-            pipeline.join()
-
-            return np.array(sink.timestamps)
-
-        # test timings for increasing number of array sizes and increasing samplerates
-        print("starting load test for fixed rate source...")
-        for shape in [None]:  # , (100,), (100, 100, 100), (1000, 1000, 1000)
-            for samplerate in [1000]: # 10, 100,
-                timestamps = _run_test_pipeline(samplerate=samplerate, shape=shape)
-                n = len(timestamps)
-                frametimes = np.diff(timestamps)
-                mean, sd, t_min, t_max = frametimes.mean(), frametimes.std(), frametimes.min(), frametimes.max()
-
-                print(f"shape={shape}, rate={samplerate}, n_rec={n}, t_mean={mean} (SD={sd}, range=[{t_min};{t_max}])")
-                pass
-                # TODO: asserts
-
-        self.assertTrue(False)
-
-
